lowered default rows per software id to 50k. log culling now happens by date which is configurable with a retention period, it also doesn't delete until a tolerance over the max entries and then deletes them all at once

+++ src/Rrs.Logging.SqlServer/LoggerQueries.cs
@@ -7,11 +7,15 @@ namespace Rrs.Logging.SqlServer
     {
         private readonly string _logTable;
         private readonly int _maxEntries;
+        private readonly int _retentionDays;
+        private readonly double _tolerance;
 
-        public LoggerQueries(string logTable, int maxEntries)
+        public LoggerQueries(string logTable, int maxEntries, int retentionDays, double tolerance)
         {
             _logTable = logTable;
             _maxEntries = maxEntries;
+            _retentionDays = retentionDays;
+            _tolerance = tolerance;
         }
 
         public void Create(IDbConnection c, LogEntry log)
@@ -19,27 +23,36 @@ namespace Rrs.Logging.SqlServer
 	        if (log.ObjectType.Length > 250) 
 		        log.ObjectType = log.ObjectType.Substring(0, 249);
 
+			// check entries over max entries + tolerance percentage
+			// delete by date to number of retention days
+			// check max entries again in case all the logs are within that retention period
+			// delete by max entries
 	        var command = $@"
 insert into {_logTable} (SoftwareId, Level, Object, ObjectType, Created) values (@SoftwareId, @Level, @Object, @ObjectType, getdate())
 
-delete from {_logTable}
-where Id in (
-	select Id
-	from
-	(
-		select Id, row_number() over(order by fromtheashes, Id desc) n
-		from (
-			select case when Id < ident_current('{_logTable}') + 1 then 0 else 1 end fromtheashes, Id
-			from {_logTable}
-            where SoftwareId = @SoftwareId
+if (select min(Created) from {_logTable} where SoftwareId = @SoftwareId) < dateadd(day, -{_retentionDays}*{_tolerance}, getdate())
+	delete from {_logTable}
+	where	SoftwareId = @SoftwareId and 
+			Created < dateadd(day, -{_retentionDays}, getdate())
+
+if (select count(*) from {_logTable} where  SoftwareId = @SoftwareId) > {_maxEntries} * {_tolerance}
+	delete from Log
+	where Id in (
+		select Id
+		from
+		(
+			select Id, row_number() over(order by fromtheashes, Id desc) n
+			from (
+				select case when Id < ident_current('{_logTable}') + 1 then 0 else 1 end fromtheashes, Id
+				from {_logTable}
+				where SoftwareId = @SoftwareId
+			) t
 		) t
-	) t
-	where n - {_maxEntries} > 0
-) 
+		where n - {_maxEntries} > 0
+	) 
 
-if (ident_current('{_logTable}')) > 1000000000
-    dbcc checkident('{_logTable}', RESEED, 0)
-
+if ident_current('{_logTable}') > 1000000000
+	dbcc checkident('{_logTable}', RESEED, 0)
 ";
             c.Sql(command).Parameters(log).Execute();
         }
